src/backend/requirements/base.txt
asgiref==3.7.2
boto3==1.28.30
botocore==1.31.30
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.7
Django==4.1.10
django-allauth==0.54.0
dnspython==2.4.2
docutils==0.20.1
filelock==3.12.2
gunicorn==21.2.0
identify==2.5.26
idna==3.4
jmespath==1.0.1
mysqlclient==2.2.0
nodeenv==1.8.0
oauthlib==3.2.2
openedx_course_structure @ git+https://github.com/epfl-cede/openedx-course-structure.git@409b5d1b6001a61800c0f797b7e18d12c99a51d0
packaging==23.1
platformdirs==3.10.0
pre-commit==3.3.3
psutil==5.9.5
pycparser==2.21
PyJWT==2.8.0
pymongo==4.4.1
python-dateutil==2.8.2
python-decouple==3.8
python-gnupg==0.5.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
Unipath==1.1
urllib3==1.26.16
virtualenv==20.24.3
